third_party/WebKit/LayoutTests/inspector-enabled/console/console-uncaught-promise-no-inspector-expected.txt

CONSOLE ERROR: line 7: Uncaught (in promise) Error: err1
CONSOLE ERROR: line 25: Uncaught (in promise) Error: err2
CONSOLE ERROR: line 42: Uncaught (in promise) NotFoundError: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
Tests that uncaught promise rejection messages have line numbers when the inspector is closed and stack traces are not collected.

console-uncaught-pro…no-inspector.html:7 Uncaught (in promise) Error: err1
console-uncaught-pro…o-inspector.html:25 Uncaught (in promise) Error: err2
console-uncaught-pro…o-inspector.html:42 Uncaught (in promise) DOMException: Failed to execute 'removeChild' on 'Node': The node to be removed is not a child of this node.
